requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
streamlit==1.31.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
numpy==1.26.4
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
opencv-python-headless==4.9.0.80
albumentations==1.3.1
scikit-learn==1.4.0
python-multipart==0.0.9
requests==2.31.0
Pillow==10.2.0
tqdm==4.66.2
timm==0.9.12
grad-cam==1.5.0
onnx==1.15.0
onnxruntime==1.17.0
